Fix TestHelloWorldAppExample
The test should be checking that the HTTP response *contains* the text "Hello, World" rather than checking for exact equality. It turns out the book itself had the right code in it, but when copying the test code to this repo, I somehow got an old versison of this one test.